chore: create new file for tokenOptions
